tests/wpt/web-platform-tests/referrer-policy/unsafe-url/http-csp/same-origin/http-http/img-tag/generic.no-redirect.http.html

<!DOCTYPE html>
<!-- DO NOT EDIT! Generated by referrer-policy/generic/tools/generate.py using referrer-policy/generic/template/test.release.html.template. -->
<html>
  <head>
    <title>Referrer-Policy: Referrer Policy is set to 'unsafe-url'</title>
    <meta name="description" content="Check that all sub-resources get the stripped referrer URL.">
    <!-- No meta: CSP delivered via HTTP headers. -->
    <link rel="author" title="Kristijan Burnik" href="burnik@chromium.org">
    <link rel="help" href="https://w3c.github.io/webappsec/specs/referrer-policy/#referrer-policy-state-unsafe-url">
    <meta name="assert" content="The referrer URL is stripped-referrer when a
                                 document served over http requires an http
                                 sub-resource via img-tag using the http-csp
                                 delivery method with no-redirect and when
                                 the target request is same-origin.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- TODO(kristijanburnik): Minify and merge both: -->
    <script src="/referrer-policy/generic/common.js"></script>
    <script src="/referrer-policy/generic/referrer-policy-test-case.js?pipe=sub"></script>
  </head>
  <body>
    <script>
      ReferrerPolicyTestCase(
        {
          "referrer_policy": "unsafe-url",
          "delivery_method": "http-csp",
          "redirection": "no-redirect",
          "origin": "same-origin",
          "source_protocol": "http",
          "target_protocol": "http",
          "subresource": "img-tag",
          "subresource_path": "/referrer-policy/generic/subresource/image.py",
          "referrer_url": "stripped-referrer"
        },
        document.querySelector("meta[name=assert]").content,
        new SanityChecker()
      ).start();
      </script>
    <div id="log"></div>
  </body>
</html>
