third_party/WebKit/LayoutTests/platform/win/external/wpt/referrer-policy/origin/attr-referrer/same-origin/http-http/img-tag/generic.no-redirect.http-expected.txt

This is a testharness.js-based test.
PASS The referrer URL is origin when a
                                 document served over http requires an http
                                 sub-resource via img-tag using the attr-referrer
                                 delivery method with no-redirect and when
                                 the target request is same-origin. 
Harness: the test ran to completion.
